requirements.txt
# requirements.txt

protobuf
pyserial
cobs
